add .env file and fix liner.vue bog
