+++ components/Input/Text/Area/Liner.vue
@@ -1,207 +1,54 @@
 <template>
-    <div class="flex w-full h-1/6 p-2 gap-2">
-      <div
-        class="w-1/2 rounded-[3px] group relative bg-secondary text-base-content hover:bg-none"
+  <div class="card flex flex-row text-sm overflow-scroll rounded-[3px]" :class="[h, w]" dir="ltr">
+    <div class=" number-card py-2 border border-r-0 bg-base-200 h-fit min-h-full" ref="numberCardDiv">
+      <span
+        v-for="line in lines"
+        class="text-base-content flex flex-row justify-center w-6"
       >
-        <div
-          class="h-full w-full rounded-[3px] flex justify-center items-center gap-3 [&>svg]:fill-base-content group-hover:hidden"
-        >
-          <span class="[&>svg]:fill-base-content">
-            <svg
-              width="18"
-              height="18"
-              viewBox="0 0 18 18"
-              fill="none"
-              xmlns="http://www.w3.org/2000/svg"
-            >
-              <path
-                d="M16.0234 17.05L10.2484 11.3C9.74844 11.7167 9.17344 12.0417 8.52344 12.275C7.87344 12.5083 7.20677 12.625 6.52344 12.625C4.80677 12.625 3.35677 12.0333 2.17344 10.85C0.990104 9.66667 0.398438 8.21667 0.398438 6.5C0.398438 4.8 0.990104 3.354 2.17344 2.162C3.35677 0.970667 4.80677 0.375 6.52344 0.375C8.22344 0.375 9.6651 0.966667 10.8484 2.15C12.0318 3.33333 12.6234 4.78333 12.6234 6.5C12.6234 7.21667 12.5068 7.9 12.2734 8.55C12.0401 9.2 11.7234 9.76667 11.3234 10.25L17.0984 16.025C17.2318 16.1583 17.2984 16.325 17.2984 16.525C17.2984 16.725 17.2234 16.9 17.0734 17.05C16.9234 17.2 16.7444 17.275 16.5364 17.275C16.3278 17.275 16.1568 17.2 16.0234 17.05ZM6.52344 11.125C7.80677 11.125 8.89444 10.675 9.78644 9.775C10.6778 8.875 11.1234 7.78333 11.1234 6.5C11.1234 5.21667 10.6778 4.125 9.78644 3.225C8.89444 2.325 7.80677 1.875 6.52344 1.875C5.22344 1.875 4.12777 2.325 3.23644 3.225C2.34444 4.125 1.89844 5.21667 1.89844 6.5C1.89844 7.78333 2.34444 8.875 3.23644 9.775C4.12777 10.675 5.22344 11.125 6.52344 11.125Z"
-              />
-            </svg>
-          </span>
-          <span class="text-base-content">
-            {{ cn.by_route(`${department_section}/seo/title`) }}
-          </span>
-        </div>
-        <div
-          class="h-full w-full bg-base-100 rounded-[3px] justify-center items-center hidden group-hover:flex"
-        >
-          <div class="flex w-full h-full p-2 gap-2">
-            <NuxtLink
-              to="/keyword-research"
-              class="bg-secondary text-base-content flex justify-center items-center cursor-pointer w-1/2 h-full rounded-[3px] group relative gap-3"
-            >
-              <span class="[&>svg]:fill-base-content">
-                <svg
-                  width="18"
-                  height="18"
-                  viewBox="0 0 18 18"
-                  fill="none"
-                  xmlns="http://www.w3.org/2000/svg"
-                >
-                  <path
-                    d="M5.62922 18C4.27153 17.9991 2.95998 17.5072 1.93656 16.6151C0.913149 15.7229 0.246949 14.4907 0.0608936 13.1459C-0.125161 11.801 0.181485 10.4342 0.924248 9.29769C1.66701 8.1612 2.79576 7.3317 4.10222 6.96227C5.05182 6.69033 6.05694 6.6779 7.01297 6.92626L12.944 0.996764C13.2588 0.679727 13.6335 0.428365 14.0463 0.257261C14.459 0.0861575 14.9017 -0.00128235 15.3485 1.42104e-05C16.0516 0.000610035 16.7258 0.280208 17.2231 0.777425C17.7203 1.27464 17.9999 1.94884 18.0005 2.65201C18.0016 3.09886 17.9142 3.5415 17.7433 3.95435C17.5723 4.36719 17.3212 4.74206 17.0045 5.05726L16.5005 5.56126C16.2185 5.84134 15.8374 5.99899 15.44 6.00001H14.2505V6.75001C14.2505 7.14784 14.0924 7.52937 13.8111 7.81067C13.5298 8.09198 13.1483 8.25002 12.7505 8.25002H12.0005V9.43952C12.001 9.63657 11.9624 9.83177 11.887 10.0138C11.8116 10.1958 11.7007 10.3611 11.561 10.5L11.0735 10.9875C11.3232 11.9431 11.311 12.9483 11.0382 13.8975C10.7313 14.979 10.1085 15.9443 9.24958 16.6697C8.3907 17.3951 7.33484 17.8476 6.21722 17.9693C6.02188 17.9896 5.82562 17.9998 5.62922 18ZM5.62922 8.25002C4.86133 8.24927 4.10848 8.46288 3.45541 8.8668C2.80234 9.27073 2.27498 9.84892 1.9327 10.5363C1.59041 11.2237 1.44679 11.993 1.518 12.7575C1.58921 13.5221 1.87242 14.2516 2.33576 14.864C2.7991 15.4763 3.42416 15.9472 4.14059 16.2236C4.85702 16.5 5.63637 16.5709 6.39091 16.4283C7.14546 16.2858 7.84524 15.9355 8.41148 15.4168C8.97772 14.8981 9.38794 14.2317 9.59597 13.4925C9.82997 12.6799 9.80374 11.8145 9.52097 11.0175C9.47482 10.8844 9.46709 10.741 9.49866 10.6037C9.53023 10.4665 9.59982 10.3408 9.69947 10.2413L10.5005 9.43952V8.25002C10.5005 7.85219 10.6585 7.47066 10.9398 7.18935C11.2211 6.90805 11.6026 6.75001 12.0005 6.75001H12.7505V6.00001C12.7505 5.60219 12.9085 5.22066 13.1898 4.93935C13.4711 4.65805 13.8526 4.50001 14.2505 4.50001H15.44L15.944 3.99601C16.1209 3.81988 16.2612 3.61041 16.3568 3.37972C16.4523 3.14903 16.5011 2.9017 16.5005 2.65201C16.5003 2.34668 16.3789 2.05389 16.1631 1.83791C15.9473 1.62193 15.6546 1.50041 15.3492 1.50001C15.0994 1.4994 14.8519 1.54833 14.621 1.64398C14.3902 1.73963 14.1807 1.8801 14.0045 2.05726L7.75622 8.30476C7.65651 8.40443 7.53072 8.47396 7.39329 8.5054C7.25586 8.53684 7.11235 8.52891 6.97922 8.48252C6.54427 8.32938 6.08659 8.25077 5.62547 8.25002H5.62922ZM3.75047 13.5C3.75047 13.6484 3.79446 13.7934 3.87687 13.9167C3.95928 14.04 4.07641 14.1362 4.21346 14.1929C4.3505 14.2497 4.5013 14.2645 4.64679 14.2356C4.79227 14.2067 4.92591 14.1352 5.0308 14.0303C5.13569 13.9255 5.20712 13.7918 5.23606 13.6463C5.265 13.5008 5.25015 13.35 5.19338 13.213C5.13662 13.076 5.04049 12.9588 4.91715 12.8764C4.79381 12.794 4.64881 12.75 4.50047 12.75C4.30156 12.75 4.11079 12.829 3.97014 12.9697C3.82949 13.1103 3.75047 13.3011 3.75047 13.5Z"
-                  />
-                </svg>
-              </span>
-              <span class="text-base-content">
-                {{ cn.by_route(`${department_section}/seo/fields/keyword-research`) }}
-              </span>
-            </NuxtLink>
-            <NuxtLink
-              to="/google-title-builder"
-              class="bg-secondary text-base-content flex justify-center items-center cursor-pointer w-1/2 h-full rounded-[3px] group relative gap-3"
-            >
-              <span class="[&>svg]:fill-base-content">
-                <svg
-                  width="18"
-                  height="18"
-                  viewBox="0 0 18 18"
-                  fill="none"
-                  xmlns="http://www.w3.org/2000/svg"
-                >
-                  <g clip-path="url(#clip0_6758_4482)">
-                    <path
-                      d="M13.992 0.695113L4.84802 9.83911C4.49877 10.1865 4.2219 10.5997 4.03343 11.0547C3.84495 11.5098 3.74862 11.9978 3.75001 12.4904V13.4976C3.75001 13.6965 3.82903 13.8873 3.96968 14.0279C4.11034 14.1686 4.3011 14.2476 4.50001 14.2476H5.50726C5.99984 14.249 6.48779 14.1527 6.94288 13.9642C7.39797 13.7757 7.81116 13.4989 8.15851 13.1496L17.3025 4.00561C17.7408 3.56624 17.987 2.97097 17.987 2.35036C17.987 1.72976 17.7408 1.13448 17.3025 0.695113C16.8568 0.269033 16.2639 0.03125 15.6473 0.03125C15.0306 0.03125 14.4378 0.269033 13.992 0.695113ZM16.242 2.94511L7.09801 12.0891C6.67511 12.5094 6.10352 12.746 5.50726 12.7476H5.25001V12.4904C5.25158 11.8941 5.48819 11.3225 5.90851 10.8996L15.0525 1.75561C15.2127 1.60258 15.4257 1.51718 15.6473 1.51718C15.8688 1.51718 16.0818 1.60258 16.242 1.75561C16.3995 1.9135 16.4879 2.12738 16.4879 2.35036C16.4879 2.57334 16.3995 2.78723 16.242 2.94511Z"
-                    />
-                    <path
-                      d="M17.25 6.73425C17.0511 6.73425 16.8603 6.81327 16.7197 6.95392C16.579 7.09457 16.5 7.28534 16.5 7.48425V11.25H13.5C12.9033 11.25 12.331 11.4871 11.909 11.909C11.4871 12.331 11.25 12.9033 11.25 13.5V16.5H3.75C3.15326 16.5 2.58097 16.2629 2.15901 15.841C1.73705 15.419 1.5 14.8467 1.5 14.25V3.75C1.5 3.15326 1.73705 2.58097 2.15901 2.15901C2.58097 1.73705 3.15326 1.5 3.75 1.5H10.5315C10.7304 1.5 10.9212 1.42098 11.0618 1.28033C11.2025 1.13968 11.2815 0.948912 11.2815 0.75C11.2815 0.551088 11.2025 0.360322 11.0618 0.21967C10.9212 0.0790176 10.7304 0 10.5315 0L3.75 0C2.7558 0.00119089 1.80267 0.396662 1.09967 1.09967C0.396662 1.80267 0.00119089 2.7558 0 3.75L0 14.25C0.00119089 15.2442 0.396662 16.1973 1.09967 16.9003C1.80267 17.6033 2.7558 17.9988 3.75 18H12.2573C12.7499 18.0014 13.238 17.9051 13.6932 17.7166C14.1484 17.5281 14.5618 17.2513 14.9093 16.902L16.9013 14.9085C17.2506 14.5612 17.5275 14.148 17.7161 13.6929C17.9047 13.2378 18.0012 12.7499 18 12.2573V7.48425C18 7.28534 17.921 7.09457 17.7803 6.95392C17.6397 6.81327 17.4489 6.73425 17.25 6.73425ZM13.8488 15.8415C13.5473 16.1423 13.166 16.3505 12.75 16.4415V13.5C12.75 13.3011 12.829 13.1103 12.9697 12.9697C13.1103 12.829 13.3011 12.75 13.5 12.75H16.4438C16.351 13.1651 16.143 13.5457 15.8438 13.848L13.8488 15.8415Z"
-                    />
-                  </g>
-                  <defs>
-                    <clipPath id="clip0_6758_4482">
-                      <rect width="18" height="18" fill="white" />
-                    </clipPath>
-                  </defs>
-                </svg>
-              </span>
-              <span class="text-base-content">
-                {{ cn.by_route(`${department_section}/seo/fields/google-title-builder`) }}
-              </span>
-            </NuxtLink>
-          </div>
-        </div>
-      </div>
-      <div
-        class="w-1/2 rounded-[3px] group relative bg-secondary text-base-content hover:bg-none"
-      >
-        <div
-          class="h-full w-full rounded-[3px] flex justify-center items-center gap-3 [&>svg]:fill-base-content group-hover:hidden"
-        >
-          <span class="[&>svg]:fill-base-content">
-            <svg
-              width="18"
-              height="18"
-              viewBox="0 0 18 18"
-              fill="none"
-              xmlns="http://www.w3.org/2000/svg"
-            >
-              <g clip-path="url(#clip0_84_933)">
-                <path
-                  d="M17.1202 0.880517C16.7563 0.51328 16.303 0.247182 15.8049 0.108436C15.3069 -0.0303103 14.7813 -0.0369168 14.2799 0.089267C10.3931 1.10968 6.83344 3.11109 3.94194 5.90177C3.37314 6.47401 2.92958 7.15831 2.63946 7.91118C2.34933 8.66406 2.21898 9.46906 2.25669 10.275C2.31041 11.445 2.72107 12.5704 3.43344 13.5L0.219689 16.7198C0.148056 16.789 0.0909195 16.8717 0.0516128 16.9632C0.012306 17.0547 -0.00838366 17.1531 -0.00924903 17.2527C-0.0101144 17.3523 0.0088619 17.4511 0.0465726 17.5432C0.0842832 17.6354 0.139973 17.7191 0.210393 17.7896C0.280812 17.86 0.364551 17.9157 0.456723 17.9534C0.548896 17.9911 0.647655 18.0101 0.74724 18.0092C0.846824 18.0083 0.945239 17.9877 1.03674 17.9483C1.12825 17.909 1.211 17.8519 1.28019 17.7803L4.49319 14.5673C5.47801 15.3362 6.69224 15.7526 7.94169 15.75C8.76848 15.738 9.58481 15.5633 10.3441 15.2358C11.1033 14.9083 11.7906 14.4344 12.3667 13.8413C15.6667 10.5413 17.5289 5.25677 17.9122 3.69752C18.0368 3.19965 18.0293 2.67789 17.8903 2.18382C17.7514 1.68976 17.486 1.24051 17.1202 0.880517ZM3.75519 10.206C3.72696 9.60807 3.8234 9.01079 4.03837 8.4521C4.25335 7.89342 4.58217 7.38556 5.00394 6.96077C5.32494 6.64202 5.65869 6.34277 5.99994 6.05027V10.9395L4.50744 12.432C4.05264 11.7754 3.79194 11.0039 3.75519 10.206ZM11.2837 12.8003C10.5694 13.5766 9.60154 14.0726 8.5542 14.1991C7.50687 14.3256 6.44877 14.0742 5.57019 13.4903L7.27944 11.781L9.31044 9.75002H13.6319C12.9526 10.8424 12.166 11.8642 11.2837 12.8003ZM16.4587 3.33527C16.0066 5.04529 15.352 6.69522 14.5087 8.25002H10.8104L13.2802 5.78027C13.3518 5.71108 13.409 5.62832 13.4483 5.53682C13.4876 5.44532 13.5083 5.3469 13.5091 5.24732C13.51 5.14773 13.491 5.04897 13.4533 4.9568C13.4156 4.86463 13.3599 4.78089 13.2895 4.71047C13.2191 4.64005 13.1353 4.58436 13.0432 4.54665C12.951 4.50894 12.8522 4.48996 12.7526 4.49083C12.6531 4.49169 12.5546 4.51238 12.4631 4.55169C12.3716 4.591 12.2889 4.64814 12.2197 4.71977L8.46969 8.46977L7.49994 9.43952V4.88402C9.6682 3.36444 12.0849 2.23476 14.6414 1.54577C14.7601 1.51589 14.882 1.50052 15.0044 1.50002C15.201 1.50044 15.3955 1.53968 15.5769 1.61549C15.7582 1.6913 15.9228 1.80218 16.0612 1.94177C16.2439 2.11872 16.3765 2.34077 16.4458 2.58547C16.5151 2.83017 16.5185 3.08883 16.4557 3.33527H16.4587Z"
-                />
-              </g>
-              <defs>
-                <clipPath id="clip0_84_933">
-                  <rect width="18" height="18" fill="white" />
-                </clipPath>
-              </defs>
-            </svg>
-          </span>
-          <span class="text-base-content">
-            {{ cn.by_route(`${department_section}/content-creation/title`) }}
-          </span>
-        </div>
-        <div
-          class="h-full w-full bg-base-100 rounded-[3px] justify-center items-center hidden group-hover:flex"
-        >
-          <div class="flex w-full h-full p-2 gap-2">
-            <NuxtLink
-              to="/title-copy-writer"
-              class="bg-secondary text-base-content flex justify-center items-center cursor-pointer w-1/2 h-full rounded-[3px] group relative gap-3"
-            >
-              <span class="[&>svg]:fill-base-content">
-                <svg
-                  width="18"
-                  height="18"
-                  viewBox="0 0 18 18"
-                  fill="none"
-                  xmlns="http://www.w3.org/2000/svg"
-                >
-                  <path
-                    d="M14.25 0H3.75C2.7558 0.00119089 1.80267 0.396661 1.09966 1.09966C0.396661 1.80267 0.00119089 2.7558 0 3.75L0 14.25C0.00119089 15.2442 0.396661 16.1973 1.09966 16.9003C1.80267 17.6033 2.7558 17.9988 3.75 18H14.25C15.2442 17.9988 16.1973 17.6033 16.9003 16.9003C17.6033 16.1973 17.9988 15.2442 18 14.25V3.75C17.9988 2.7558 17.6033 1.80267 16.9003 1.09966C16.1973 0.396661 15.2442 0.00119089 14.25 0V0ZM16.5 14.25C16.5 14.8467 16.2629 15.419 15.841 15.841C15.419 16.2629 14.8467 16.5 14.25 16.5H3.75C3.15326 16.5 2.58097 16.2629 2.15901 15.841C1.73705 15.419 1.5 14.8467 1.5 14.25V3.75C1.5 3.15326 1.73705 2.58097 2.15901 2.15901C2.58097 1.73705 3.15326 1.5 3.75 1.5H14.25C14.8467 1.5 15.419 1.73705 15.841 2.15901C16.2629 2.58097 16.5 3.15326 16.5 3.75V14.25ZM13.5 6.75C13.5 6.94891 13.421 7.13968 13.2803 7.28033C13.1397 7.42098 12.9489 7.5 12.75 7.5C12.5511 7.5 12.3603 7.42098 12.2197 7.28033C12.079 7.13968 12 6.94891 12 6.75C12 6.55109 11.921 6.36032 11.7803 6.21967C11.6397 6.07902 11.4489 6 11.25 6H9.75V12H10.5C10.6989 12 10.8897 12.079 11.0303 12.2197C11.171 12.3603 11.25 12.5511 11.25 12.75C11.25 12.9489 11.171 13.1397 11.0303 13.2803C10.8897 13.421 10.6989 13.5 10.5 13.5H7.5C7.30109 13.5 7.11032 13.421 6.96967 13.2803C6.82902 13.1397 6.75 12.9489 6.75 12.75C6.75 12.5511 6.82902 12.3603 6.96967 12.2197C7.11032 12.079 7.30109 12 7.5 12H8.25V6H6.75C6.55109 6 6.36032 6.07902 6.21967 6.21967C6.07902 6.36032 6 6.55109 6 6.75C6 6.94891 5.92098 7.13968 5.78033 7.28033C5.63968 7.42098 5.44891 7.5 5.25 7.5C5.05109 7.5 4.86032 7.42098 4.71967 7.28033C4.57902 7.13968 4.5 6.94891 4.5 6.75C4.5 6.15326 4.73705 5.58097 5.15901 5.15901C5.58097 4.73705 6.15326 4.5 6.75 4.5H11.25C11.8467 4.5 12.419 4.73705 12.841 5.15901C13.2629 5.58097 13.5 6.15326 13.5 6.75Z"
-                  />
-                </svg>
-              </span>
-              <span class="text-base-content">
-                {{
-                  cn.by_route(
-                    `${department_section}/content-creation/fields/title-copy-writer`
-                  )
-                }}
-              </span>
-            </NuxtLink>
-            <NuxtLink
-              to="/idea-generator"
-              class="bg-secondary text-base-content flex justify-center items-center cursor-pointer w-1/2 h-full rounded-[3px] group relative gap-3"
-            >
-              <span class="[&>svg]:fill-base-content">
-                <svg
-                  width="14"
-                  height="18"
-                  viewBox="0 0 14 18"
-                  fill="none"
-                  xmlns="http://www.w3.org/2000/svg"
-                >
-                  <path
-                    d="M11.4917 1.71625C10.5762 0.898405 9.45509 0.345685 8.24891 0.117571C7.04273 -0.110543 5.79718 -0.00541525 4.64631 0.421642C3.49543 0.848699 2.48279 1.58151 1.71738 2.54123C0.951975 3.50094 0.462768 4.65121 0.302412 5.86825C0.158397 6.93538 0.273576 8.02151 0.638201 9.03471C1.00283 10.0479 1.60617 10.9583 2.39716 11.689C2.88631 12.118 3.28102 12.6438 3.55627 13.2333C3.83153 13.8228 3.98132 14.4631 3.99616 15.1135V15.2342C3.99696 15.968 4.28879 16.6715 4.80762 17.1903C5.32646 17.7091 6.02992 18.001 6.76366 18.0017H7.22866C7.96241 18.001 8.66587 17.7091 9.1847 17.1903C9.70354 16.6715 9.99537 15.968 9.99616 15.2342V14.8225C10.0074 14.2476 10.1427 13.6819 10.3928 13.1641C10.6428 12.6463 11.0018 12.1887 11.4452 11.8225C12.1641 11.1928 12.741 10.4176 13.1377 9.54811C13.5345 8.67863 13.7419 7.73479 13.7463 6.7791C13.7507 5.82341 13.5519 4.8777 13.1632 4.00461C12.7746 3.13153 12.2048 2.351 11.4917 1.71475V1.71625ZM7.22866 16.5017H6.76366C6.42762 16.5014 6.10546 16.3677 5.86784 16.1301C5.63023 15.8924 5.49656 15.5703 5.49616 15.2342C5.49616 15.2342 5.49091 15.0392 5.49016 15.0017H8.49616V15.2342C8.49576 15.5703 8.3621 15.8924 8.12448 16.1301C7.88687 16.3677 7.5647 16.5014 7.22866 16.5017ZM10.4537 10.696C9.59973 11.4367 8.98304 12.4126 8.68066 13.5017H7.74616V8.11375C8.18334 7.95919 8.56209 7.67332 8.8306 7.29527C9.0991 6.91722 9.24424 6.46544 9.24616 6.00175C9.24616 5.80283 9.16714 5.61207 9.02649 5.47142C8.88584 5.33077 8.69507 5.25175 8.49616 5.25175C8.29725 5.25175 8.10648 5.33077 7.96583 5.47142C7.82518 5.61207 7.74616 5.80283 7.74616 6.00175C7.74616 6.20066 7.66714 6.39143 7.52649 6.53208C7.38584 6.67273 7.19507 6.75175 6.99616 6.75175C6.79725 6.75175 6.60648 6.67273 6.46583 6.53208C6.32518 6.39143 6.24616 6.20066 6.24616 6.00175C6.24616 5.80283 6.16714 5.61207 6.02649 5.47142C5.88584 5.33077 5.69507 5.25175 5.49616 5.25175C5.29725 5.25175 5.10648 5.33077 4.96583 5.47142C4.82518 5.61207 4.74616 5.80283 4.74616 6.00175C4.74808 6.46544 4.89323 6.91722 5.16173 7.29527C5.43023 7.67332 5.80898 7.95919 6.24616 8.11375V13.5017H5.25466C4.91641 12.3837 4.282 11.3779 3.41866 10.591C2.68588 9.90819 2.16371 9.03002 1.91376 8.06011C1.66382 7.0902 1.69655 6.06904 2.00807 5.11712C2.31959 4.1652 2.89692 3.32226 3.67191 2.68777C4.44691 2.05328 5.38723 1.65372 6.38191 1.53625C6.58811 1.51323 6.79543 1.50171 7.00291 1.50175C8.29109 1.49792 9.53467 1.97322 10.4919 2.83525C11.0467 3.32995 11.49 3.93691 11.7924 4.61592C12.0948 5.29492 12.2494 6.03045 12.246 6.77374C12.2426 7.51703 12.0812 8.25111 11.7726 8.92731C11.464 9.60352 11.0152 10.2064 10.4559 10.696H10.4537Z"
-                  />
-                </svg>
-              </span>
-              <span class="text-base-content">
-                {{
-                  cn.by_route(
-                    `${department_section}/content-creation/fields/idea-generator`
-                  )
-                }}
-              </span>
-            </NuxtLink>
-          </div>
-        </div>
-      </div>
+        {{ line }}
+      </span>
     </div>
-    <div class="flex w-full px-2 h-[10%] justify-center items-center">
-      <Copy content="استارتر" class="w-full h-full">
-        <NuxtLink
-          to="/buy-subscription"
-          class="rounded-[3px] w-full h-full flex justify-around items-center bg-gradient-to-r from-[#1444b0] to-primary text-base-100"
-        >
-          <span> همین الان با 50% تخفیف شروع کنید.</span>
-          <span
-            >کد تخفیف:
-            <span>استارتر</span>
-          </span>
-        </NuxtLink>
-      </Copy>
-    </div>
-    <div class="w-full flex flex-row justify-center items-center">
-      <InputTextAreaLiner v-model="text"> </InputTextAreaLiner>
-    </div>
-    <div>
-      {{ text }}
-    </div>
-  </template>
-  
-  <script setup>
-  import Config from "../composables/Config";
-  import Auth from "../middlewares/Auth";
-  
-  definePageMeta({
-    middleware: [Auth],
-  });
-  
-  const text = ref("")
-  const cn = new Config();
-  const department_section = "layouts/default/navbar/right/department";
-  </script>
-  
+    <textarea
+    ref="textAreaDiv"
+      :value="modelValue"
+      @input="$emit('update:modelValue', $event.target.value)"
+      @keyup="line_counter($event)"
+      class="resize-none overflow-hidden border border-b-[2px] focus:border-b-primary p-2 w-full"
+      :class="class"
+    ></textarea>
+  </div>
+</template>
+
+<script setup>
+const lines = ref(1);
+const numberCardDiv = ref(null)
+const textAreaDiv = ref(null)
+
+const props = defineProps({
+  class: {
+    type: String,
+    default: "",
+  },
+  h: {
+    type: String,
+    default: "h-44",
+  },
+  w: {
+    type: String,
+    default: "w-1/2",
+  },
+  modelValue:{}
+});
+defineEmits(["update:modelValue"]);
+function line_counter(event) {
+  lines.value = event.target.value.split("\n").length;
+  // event.target.style.height = "100%";
+  if (lines.value > 7) {
+    numberCardDiv.value.style.height = "1000px";
+    textAreaDiv.value.style.height = "1000px";
+  }else {
+    numberCardDiv.value.style.height = "170px";
+    textAreaDiv.value.style.height = "170px";
+  }
+}
+</script>
